fix(web): resume checkout from join
