+++ apps/web/app/(dashboard)/biomarkers/[biomarkerId]/page.tsx
@@ -3,6 +3,7 @@ import { notFound, redirect } from "next/navigation";
 
 import { BiomarkerOverview } from "@/src/components/biomarkers/biomarker-detail/biomarker-overview";
 import { resolveHealthCommonsBiomarkerOverview } from "@/src/lib/health-commons/biomarker-projections";
+import { getHostedDashboardPageAuthSnapshot } from "@/src/lib/hosted-onboarding/page-auth";
 import { createMurphPageMetadata } from "@/src/lib/site-metadata";
 
 export async function generateMetadata({
@@ -31,6 +32,7 @@ export default async function BiomarkerOverviewPage({
 }: {
   params: Promise<{ biomarkerId: string }>;
 }) {
+  await getHostedDashboardPageAuthSnapshot();
   const { biomarkerId } = await params;
   const biomarker = resolveHealthCommonsBiomarkerOverview(biomarkerId);
