+++ apps/web/app/(dashboard)/experiments/page.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 
 import { listHealthCommonsExperimentBrowseProtocols } from "@/src/lib/health-commons/experiment-browse";
+import { getHostedDashboardPageAuthSnapshot } from "@/src/lib/hosted-onboarding/page-auth";
 import { createMurphPageMetadata } from "@/src/lib/site-metadata";
 import { ExperimentsPageClient } from "./experiments-page-client";
 
@@ -10,7 +11,8 @@ export const metadata: Metadata = createMurphPageMetadata({
     "Browse evidence-backed health experiments and compare what changes against your own baseline.",
 });
 
-export default function ExperimentsPage() {
+export default async function ExperimentsPage() {
+  await getHostedDashboardPageAuthSnapshot();
   const protocols = listHealthCommonsExperimentBrowseProtocols();
 
   return <ExperimentsPageClient protocols={protocols} />;
