+++ apps/web/app/(dashboard)/home/page.tsx
@@ -32,7 +32,7 @@ import { shouldShowHomeDeviceSyncStep } from "@/src/lib/device-sync/home-onboard
 import { listHealthCommonsExperimentBrowseProtocols } from "@/src/lib/health-commons/experiment-browse";
 import { resolveHostedAiUsageGate } from "@/src/lib/hosted-execution/usage-allowance";
 import { readHostedMemberStripeBillingRef } from "@/src/lib/hosted-onboarding/hosted-member-billing-store";
-import { getHostedPageAuthSnapshot } from "@/src/lib/hosted-onboarding/page-auth";
+import { getHostedDashboardPageAuthSnapshot } from "@/src/lib/hosted-onboarding/page-auth";
 import { getPrisma } from "@/src/lib/prisma";
 import { createMurphPageMetadata } from "@/src/lib/site-metadata";
 
@@ -56,7 +56,7 @@ export default async function HomePage({
   const resolvedSearchParams = searchParams ? await searchParams : {};
   const showInitialVisitDialog =
     readFirstSearchParamValue(resolvedSearchParams.initialVisit) === "true";
-  const auth = await getHostedPageAuthSnapshot();
+  const auth = await getHostedDashboardPageAuthSnapshot();
   const member = auth.authenticatedMember;
 
   const prisma = getPrisma();
