+++ apps/web/app/(dashboard)/ops/runtime-maintenance/page.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 
 import { requireHostedOpsPageAccess } from "@/src/lib/hosted-ops/access";
 import { readHostedRuntimeMaintenanceOverview } from "@/src/lib/hosted-ops/runtime-maintenance";
+import { getHostedDashboardPageAuthSnapshot } from "@/src/lib/hosted-onboarding/page-auth";
 
 import { RuntimeMaintenanceClient } from "./runtime-maintenance-client";
 
@@ -18,6 +19,7 @@ export const metadata: Metadata = {
 };
 
 export default async function RuntimeMaintenanceOpsPage() {
+  await getHostedDashboardPageAuthSnapshot();
   await requireHostedOpsPageAccess();
   const overview = await readHostedRuntimeMaintenanceOverview();
