+++ apps/web/src/lib/hosted-onboarding/sidebar-auth.ts
@@ -1,11 +1,9 @@
 export interface HostedSidebarAuthSnapshot {
   authenticated: boolean;
   label: string | null;
-  requiresDashboardRecovery: boolean;
 }
 
 export const anonymousHostedSidebarAuthSnapshot: HostedSidebarAuthSnapshot = {
   authenticated: false,
   label: null,
-  requiresDashboardRecovery: false,
 };
