+++ apps/web/test/connect-page.test.ts
@@ -87,6 +87,7 @@ vi.mock("@/src/lib/device-sync/settings-service", () => ({
 
 vi.mock("@/src/lib/hosted-onboarding/page-auth", () => ({
   getHostedPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
+  getHostedDashboardPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
 }));
 
 vi.mock("@/src/components/murph/hosted-murph-contact-action", () => ({
