+++ apps/web/test/dashboard-sidebar.test.ts
@@ -238,7 +238,6 @@ test("Sidebar ignores supplied signed-in user labels in the account trigger", ()
       initialAuth: {
         authenticated: true,
         label: "initial@example.com",
-        requiresDashboardRecovery: false,
       },
     }),
   );
@@ -258,7 +257,6 @@ test("Sidebar renders signed-in account controls without a visible fallback labe
       initialAuth: {
         authenticated: true,
         label: null,
-        requiresDashboardRecovery: false,
       },
     }),
   );
@@ -286,7 +284,6 @@ test("Sidebar surfaces a visible error when sign out fails", async () => {
       initialAuth: {
         authenticated: true,
         label: null,
-        requiresDashboardRecovery: false,
       },
     }),
     { requireButton: false },
@@ -354,7 +351,6 @@ test("Sidebar keeps Settings out of the primary navigation", () => {
       initialAuth: {
         authenticated: true,
         label: null,
-        requiresDashboardRecovery: false,
       },
     }),
   );
@@ -410,7 +406,6 @@ test("Sidebar uses initial server app-session auth", () => {
       initialAuth: {
         authenticated: true,
         label: "initial@example.com",
-        requiresDashboardRecovery: false,
       },
     }),
   );
@@ -429,7 +424,6 @@ test("Sidebar does not render a hardcoded wearable connection status", () => {
       initialAuth: {
         authenticated: true,
         label: null,
-        requiresDashboardRecovery: false,
       },
     }),
   );
