+++ apps/web/test/join-page.test.ts
@@ -77,6 +77,7 @@ vi.mock("@/src/components/hosted-onboarding/join-invite-success-client", () => (
 
 vi.mock("@/src/lib/hosted-onboarding/page-auth", () => ({
   getHostedPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
+  getHostedDashboardPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
 }));
 
 vi.mock("@/src/lib/hosted-onboarding/hosted-session", () => ({
