+++ apps/web/test/layout.test.ts
@@ -9,7 +9,6 @@ const mocks = vi.hoisted(() => ({
   getHostedSidebarAuthSnapshot: vi.fn(async () => ({
     authenticated: false,
     label: null,
-    requiresDashboardRecovery: false,
   })),
 }));
