+++ apps/web/test/not-found.test.ts
@@ -31,6 +31,7 @@ vi.mock("next/link", () => ({
 
 vi.mock("@/src/lib/hosted-onboarding/page-auth", () => ({
   getHostedPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
+  getHostedDashboardPageAuthSnapshot: mocks.getHostedPageAuthSnapshot,
 }));
 
 beforeEach(() => {
